Move Styled Components Registry
