+++ jobs/Frontend/app/layout.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from "next";
-import StyledComponentsRegistry from "@/lib/registry";
+import StyledComponentsRegistry from "@/styles/registry";
 import GlobalStyles from "@/styles/GlobalStyles";
 import ThemeProvider from "./theme-provider";
